requirements.txt
uvicorn
fastapi 
transformers
torch
Pillow
python-multipart
celery
redis